ZON-5635: split image and imageroup handling
while the first is a straightforward http fetch that creates a local
image, creating an imagegroup requires metadata specific to the
brightcove video it pertains to, so it makes sense to move that part
into the brightcove module
