+++ core/src/zeit/content/image/tests/test_fetch.py
@@ -2,10 +2,7 @@ import pytest
 import shutil
 from os import path
 from pytest_server_fixtures.http import SimpleHTTPTestServer
-from zeit.content.image import fetch
-import zeit.cms.repository.interfaces
-import zope.component
-import zeit.content.image.testing
+from zeit.content.image import image
 
 
 @pytest.yield_fixture
@@ -23,33 +20,15 @@ def test_image_server(image_server):
 
 
 def test_fetch_remote_image(image_server):
-    local_image = fetch.get_remote_image('%s/testdata/opernball.jpg' % image_server.uri)
+    local_image = image.get_remote_image('%s/testdata/opernball.jpg' % image_server.uri)
     assert local_image.mimeType == 'image/jpeg'
 
 
 def test_fetch_remote_image_fail(image_server):
-    local_image = fetch.get_remote_image('%s/testdata/nosuchimage.jpg' % image_server.uri)
+    local_image = image.get_remote_image('%s/testdata/nosuchimage.jpg' % image_server.uri)
     assert local_image is None
 
 
 def test_fetch_remote_image_fail_url(image_server):
-    local_image = fetch.get_remote_image('nosuchhost')
+    local_image = image.get_remote_image('nosuchhost')
     assert local_image is None
-
-
-class ImageGroupTest(zeit.content.image.testing.FunctionalTestCase):
-
-    def repository(self):
-        return zope.component.getUtility(
-            zeit.cms.repository.interfaces.IRepository)
-
-    def test_image_group_from_image(self):
-        repository = self.repository()
-        local_image = zeit.content.image.testing.create_local_image('opernball.jpg')
-        group = fetch.image_group_from_image(repository, 'group', local_image)
-        assert group.master_image is not None
-
-    def test_image_group_from_none(self):
-        repository = self.repository()
-        group = fetch.image_group_from_image(repository, 'group', None)
-        assert group.master_image is None
